+++ src/adapter/newsapi/mod.rs
@@ -6,7 +6,7 @@ pub use keyword::run_news_fetcher;
 pub mod prelude {
     pub use crate::config::Config;
     pub use crate::error::Result;
-    pub use crate::model::signal::SignalInfo;
+    pub use crate::model::signal::*;
     pub const SOURCE: &str = "NewsAPI";
     pub const AGENT: &str = "Signals";
 }