+++ src/database/response.rs
@@ -2,12 +2,21 @@ use serde::Deserialize;
 use std::fmt::Debug;
 
 #[derive(Debug, Deserialize)]
-pub struct QuestDBResponse<T: Debug + 'static> {
+pub struct QuestDbResponse<T: Debug + 'static> {
+    // success/error
     pub query: String,
+    // success
     pub columns: Vec<Column>,
+    // success
     pub timestamp: usize,
+    // success
     pub dataset: Vec<T>,
+    // success
     pub count: usize,
+    // error
+    pub error: Option<String>,
+    // success
+    pub position: Option<i64>,
 }
 
 #[derive(Debug, Deserialize)]
@@ -19,48 +28,54 @@ pub struct Column {
 
 #[derive(Debug, Deserialize)]
 pub struct SignalDataRow<VALUE: Debug + 'static>(
-    String, // "id" (SYMBOL)
-    VALUE,  // "value" (DOUBLE)
-    String, // "timestamp" (TIMESTAMP as ISO 8601 string)
+    pub String, // "id" (SYMBOL)
+    pub VALUE,  // "value" (DOUBLE)
+    pub String, // "timestamp" (TIMESTAMP as ISO 8601 string)
 );
 
 #[cfg(test)]
 mod tests {
+    use crate::database::table::SignalScalarRow;
 
     #[test]
-    fn test_select_singal_scalar() {
+    fn test_select_singal_scalar_ok() {
         use super::*;
 
         // Example JSON response from QuestDB
-        let json_response = r#"
-            {
-                "query": "SELECT * FROM signal_scalar;",
-                "columns": [
-                    { "name": "id", "type": "SYMBOL" },
-                    { "name": "value", "type": "DOUBLE" },
-                    { "name": "timestamp", "type": "TIMESTAMP" }
-                ],
-                "timestamp": 2,
-                "dataset": [
-                    ["DummySource", 1.0, "2025-01-20T08:08:41.221370Z"],
-                    ["DummySource", 2.0, "2025-01-20T08:08:41.221370Z"],
-                    ["DummySource", 1.0, "2025-01-20T08:08:41.221373Z"],
-                    ["DummySource", 2.0, "2025-01-20T08:08:41.221373Z"],
-                    ["DummySource", 1.0, "2025-01-20T08:16:09.237566Z"],
-                    ["DummySource", 2.0, "2025-01-20T08:16:09.237566Z"],
-                    ["DummySource", 1.0, "2025-01-20T08:16:09.237566Z"],
-                    ["DummySource", 2.0, "2025-01-20T08:16:09.237566Z"]
-                ],
-                "count": 8
-            }
+        let json_response = r#"{
+            "query": "SELECT * FROM signal_scalar;",
+            "columns": [
+                {
+                "name": "info_id",
+                "type": "LONG"
+                },
+                {
+                "name": "value",
+                "type": "DOUBLE"
+                },
+                {
+                "name": "timestamp",
+                "type": "TIMESTAMP"
+                }
+            ],
+            "timestamp": 2,
+            "dataset": [
+                [0, 1, "2025-01-21T05:54:08.189677Z"]
+            ],
+            "count": 1
+        }    
         "#;
 
         // Deserialize JSON into Rust structs
-        let response: QuestDBResponse<SignalDataRow<f64>> = serde_json::from_str(json_response).unwrap();
+        let response: QuestDbResponse<SignalScalarRow> =
+            serde_json::from_str(json_response).unwrap();
 
         // Iterate through the dataset
         for row in &response.dataset {
-            println!("Row: id={}, value={}, timestamp={}", row.0, row.1, row.2);
+            println!(
+                "Row: info_id={}, value={}, timestamp={}",
+                row.info_id, row.value, row.timestamp
+            );
         }
     }
 }