+++ src/model/signal.rs
@@ -1,26 +1,34 @@
+use std::hash::Hash;
+
 use chrono::Utc;
 use serde::{Deserialize, Serialize};
 
 /// the main signal struct
-#[derive(Default, Debug, Clone)]
+#[derive(Default, Debug, Clone, PartialEq)]
 pub struct Signal {
-    pub id: u64,
+    pub info_id: i64,
     pub timestamp_us: i64,
     pub data: SignalData,
-    pub info_id: u64,
 }
+impl Hash for Signal {
+    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
+        self.info_id.hash(state);
+        self.timestamp_us.hash(state);
+    }
+}
+impl Eq for Signal {}
+
 impl Signal {
     pub fn dummy_scalar(value: f64) -> Self {
         Self {
-            id: 0,
+            info_id: 0,
             timestamp_us: Utc::now().timestamp_micros(),
             data: SignalData::Scalar(value),
-            info_id: 0,
         }
     }
 }
 
-#[derive(Default, Debug, Clone, Deserialize, Serialize)]
+#[derive(Default, Debug, Clone, Deserialize, Serialize, PartialEq)]
 pub enum SignalData {
     #[default]
     Simple,
@@ -31,7 +39,7 @@ pub enum SignalData {
 }
 
 /// description of SignalData
-#[derive(Default, Debug, Clone, Deserialize, Serialize)]
+#[derive(Default, Debug, Clone, Copy, Deserialize, Serialize)]
 pub enum SignalDataType {
     #[default]
     Simple,
@@ -42,8 +50,9 @@ pub enum SignalDataType {
 
 #[derive(Debug, Clone, Deserialize, Serialize)]
 pub struct SignalInfo {
-    pub id: u64,
+    pub id: i64,
     pub signal_type: String,
+    pub data_type: SignalDataType,
     pub source: String,
     pub description: String,
     /// atomic: is raw and indivisible
@@ -54,6 +63,7 @@ impl SignalInfo {
         Self {
             id: 0,
             signal_type: "dummy_index".to_string(),
+            data_type: SignalDataType::Scalar,
             source: "DummySource".to_string(),
             description: "Dummy Index".to_string(),
             is_atomic: true,